added edit function in progress delete function

+++ src/main/java/library/controllers/admin/ManageMaterialController.java
@@ -47,7 +47,7 @@ public class ManageMaterialController
 	}
 
 	@RequestMapping(value = "/manage/material/save", method = RequestMethod.POST)
-	public String save(@Valid @ModelAttribute("saveMaterial") Material material, BindingResult bindingResult, @RequestParam(value = "materialType", required = true) int materialType)
+	public String save(@Valid @ModelAttribute("material") Material material, BindingResult bindingResult, @RequestParam(value = "materialType", required = true) int materialType)
 	{
 
 
@@ -77,7 +77,7 @@ public class ManageMaterialController
 	}
 
 	@RequestMapping(value = "/manage/material/update", method = RequestMethod.POST)
-	public String update(@Valid @ModelAttribute("upMaterial") Material material, BindingResult bindingResult, @RequestParam(value = "materialType", required = true) int materialType)
+	public String update(@Valid @ModelAttribute("material") Material material, BindingResult bindingResult, @RequestParam(value = "materialType", required = true) int materialType)
 	{
 
 
@@ -91,8 +91,13 @@ public class ManageMaterialController
 
 		try
 		{
+			Material temp = materialService.getMaterialById(material.getId());
+			temp.setTitle(material.getTitle());
+			temp.setAuthor(material.getAuthor());
+			temp.setPublisher(material.getPublisher());
+			temp.setYear(material.getYear());
 			material.setCategory(materialType);
-			materialService.saveMaterial(material);
+			materialService.saveMaterial(temp);
 		}
 		catch (DataIntegrityViolationException e)
 		{
@@ -106,11 +111,28 @@ public class ManageMaterialController
 		return "redirect:/manage/material";
 	}
 	@RequestMapping(value = "/manage/material/delete", method = RequestMethod.POST)
-	public String delete(@Valid @ModelAttribute("delMaterial") Material material, BindingResult bindingResult, @RequestParam(value = "materialType", required = true) int materialType)
+	public String delete(@Valid @ModelAttribute("material") Material material, BindingResult bindingResult)
 	{
+		System.out.println(String.format("Processing user create form=%s, bindingResult=%s", material, bindingResult));
 
+		if (bindingResult.hasErrors())
+		{
+			// failed validation
+			return "manage/material";
+		}
 
-		materialService.deleteMaterial(material);
+		try
+		{
+			materialService.deleteMaterial(material);
+		}
+		catch (DataIntegrityViolationException e)
+		{
+			// probably email already exists - very rare case when multiple admins are adding same user
+			// at the same time and form validation has passed for more than one of them.
+			LOGGER.warn("Exception occurred when trying to update the material, assuming duplicate material", e);
+			bindingResult.reject("material.exist", "Material already exists");
+			return "manage/material";
+		}
 		// ok, redirect
 		return "redirect:/manage/material";
 	}
